Move functions back and use browser-compatible sha256
