+++ packages/utils/package.json
@@ -54,6 +54,8 @@
     "@babel/core": "^7.18.6",
     "@metamask/snap-types": "^0.22.0",
     "@metamask/utils": "^3.1.0",
+    "@noble/hashes": "^1.1.3",
+    "@scure/base": "^1.1.1",
     "ajv": "^8.11.0",
     "eth-rpc-errors": "^4.0.3",
     "fast-deep-equal": "^3.1.3",
