+++ packages/utils/src/snaps.ts
@@ -1,5 +1,7 @@
 import { SerializedEthereumRpcError } from 'eth-rpc-errors/dist/classes';
 import { Json } from '@metamask/utils';
+import { sha256 } from '@noble/hashes/sha256';
+import { base64 } from '@scure/base';
 import { SnapManifest } from './json-schemas';
 import { SnapId, SnapIdPrefixes, SnapValidationFailureReason } from './types';
 
@@ -151,7 +153,38 @@ export class ProgrammaticallyFixableSnapError extends Error {
   }
 }
 
+/**
+ * Calculates the Base64-encoded SHA-256 digest of a Snap source code string.
+ *
+ * @param sourceCode - The UTF-8 string source code of a Snap.
+ * @returns The Base64-encoded SHA-256 digest of the source code.
+ */
+export function getSnapSourceShasum(sourceCode: string): string {
+  return base64.encode(sha256(sourceCode));
+}
+
 export type ValidatedSnapId = `local:${string}` | `npm:${string}`;
+
+/**
+ * Checks whether the `source.shasum` property of a Snap manifest matches the
+ * shasum of a snap source code string.
+ *
+ * @param manifest - The manifest whose shasum to validate.
+ * @param sourceCode - The source code of the snap.
+ * @param errorMessage - The error message to throw if validation fails.
+ */
+export function validateSnapShasum(
+  manifest: SnapManifest,
+  sourceCode: string,
+  errorMessage = 'Invalid Snap manifest: manifest shasum does not match computed shasum.',
+): void {
+  if (manifest.source.shasum !== getSnapSourceShasum(sourceCode)) {
+    throw new ProgrammaticallyFixableSnapError(
+      errorMessage,
+      SnapValidationFailureReason.ShasumMismatch,
+    );
+  }
+}
 
 /**
  * Extracts the snap prefix from a snap ID.
